+++ models/Exercise.js
@@ -14,6 +14,7 @@ Exercise.init(
     name: {
       type: DataTypes.STRING,
       allowNull: false,
+      unique: true // this might not work
     },
     type: 
     {
